Removed unused files, made updates to code and comments
Updated the code to be fit for a showcase. Althoug there is still
a lot of TODO's in the code, these are not *ESSENTIAL* but would have
to be sorted in order to guarantee stability. However, we will not need
to do that for our grade and as we wont use this particular client after
the project, and since we due to scores on tests will only be required to
get grade 3, doing these synchronizations and redesigns for usability
just seems like a waste of time.

Changed/added some comments, general usage documentation comments.


Signed-off-by: Björn Rasmusson <bjorn.rasmusson@gmail.com>